fix(core): optimize task orderer to handle large graphs
