+++ packages/eslint-plugin-nx/src/rules/enforce-module-boundaries.ts
@@ -8,7 +8,6 @@ import {
   hasBuildExecutor,
   hasNoneOfTheseTags,
   isAbsoluteImportIntoAnotherProject,
-  checkCircularPath,
   isRelativeImportIntoAnotherProject,
   matchImportWithWildcard,
   onlyLoadChildren,
@@ -27,6 +26,7 @@ import {
   readWorkspaceJson,
 } from '@nrwl/workspace/src/core/file-utils';
 import { TargetProjectLocator } from '@nrwl/workspace/src/core/target-project-locator';
+import { checkCircularPath } from '@nrwl/workspace/src/utils/graph-utils';
 
 type Options = [
   {
